feat: add scholarship radar app workflow

+++ yolo/cvReSize.py
@@ -3,6 +3,9 @@ import os
 
 for i in range(1000):
     oldPath = f"TestData/image-{i}-old.png"
-    img = cv.resize(cv.imread(oldPath), (640, 640))
+    source = cv.imread(oldPath)
+    if source is None:
+        continue
+    img = cv.resize(source, (640, 640))
     cv.imwrite(f"TestData/image-{i}.png", img)
     os.remove(oldPath)
